services/extract_sentiment_service/requirements.txt
torch
transformers
tokenizers
huggingface_hub
numpy
confluent-kafka
flashtext
psycopg2-binary